+++ src/main/java/com/kangwon/festival/main/admin/dto/ReportResponse.java
@@ -8,9 +8,9 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 public class ReportResponse {
-    private Integer reportId;
-    private Integer reportedUserId;
-    private Integer reporterUserId;
+    private Long reportId;
+    private Long reportedUserId;
+    private Long reporterUserId;
     private String reason;
     private LocalDateTime createdDateTime;
 }
