+++ src/main/java/com/kangwon/festival/main/mypage/repository/MypageUserInfoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MypageUserInfoRepository extends JpaRepository<UserInfo, Integer> {
+public interface MypageUserInfoRepository extends JpaRepository<UserInfo, Long> {
     Optional<UserInfo> findByUserNickname(String nickname);
 }
