+++ src/main/java/com/kangwon/festival/main/mypage/service/MyPageService.java
@@ -28,7 +28,7 @@ public class MyPageService {
     private final MypageUserBlockRepository userBlockRepository;
 
     @MethodDescription(description = "닉네임을 변경합니다.")
-    public void updateNickname(int userId, String newNickname){
+    public void updateNickname(Long userId, String newNickname){
         UserInfo user = userInfoRepository.findById(userId)
                 .orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
 
@@ -47,7 +47,7 @@ public class MyPageService {
 
     @MethodDescription(description = "회원 탈퇴 처리합니다.")
     @Transactional
-    public void deleteUser(int userId) {
+    public void deleteUser(Long userId) {
         UserInfo user = userInfoRepository.findById(userId)
                 .orElseThrow(() -> new ServiceException(CAN_NOT_FIND_USER));
 
@@ -61,7 +61,7 @@ public class MyPageService {
     }
 
     @MethodDescription(description = "내 방명록을 조회합니다.")
-    public List<GuestbookResponse> getMyGuestbook(int userId) {
+    public List<GuestbookResponse> getMyGuestbook(Long userId) {
         UserInfo user = userInfoRepository.findById(userId)
                 .orElseThrow(() -> new ServiceException(CAN_NOT_FIND_USER));
 
@@ -71,7 +71,7 @@ public class MyPageService {
     }
 
     @MethodDescription(description = "내가 차단한 사용자 목록을 조회합니다.")
-    public List<UserBlockResponse> getBlockedUsers(int userId) {
+    public List<UserBlockResponse> getBlockedUsers(Long userId) {
         UserInfo user = userInfoRepository.findById(userId)
                 .orElseThrow(() -> new ServiceException(CAN_NOT_FIND_USER));
 
@@ -82,7 +82,7 @@ public class MyPageService {
 
     @MethodDescription(description = "특정 사용자의 차단을 해제합니다.")
     @Transactional
-    public void unblockUser(int blockerId, int blockedId) {
+    public void unblockUser(Long blockerId, Long blockedId) {
         UserInfo blocker = userInfoRepository.findById(blockerId)
                 .orElseThrow(() -> new ServiceException(CAN_NOT_FIND_BLOCKER_USER));
         UserInfo blocked = userInfoRepository.findById(blockedId)
